Вынести часть логики из BaseTrackersVC в хелперы
