+++ Tracker/ScheduleVC/ScheduleViewController.swift
@@ -69,8 +69,8 @@ final class ScheduleViewController: BaseTrackerViewController {
         
         cell.toggleSwitch.addTarget(self, action: #selector(switchChanged), for: .valueChanged)
         
-        configureBaseCell(cell, at: indexPath, totalRows: DayOfTheWeek.allCases.count)
-        configureSeparator(cell, isLastRow: indexPath.row == DayOfTheWeek.allCases.count - 1)
+        ConfigureTableViewCellsHelper.configureBaseCell(cell, at: indexPath, totalRows: DayOfTheWeek.allCases.count)
+        ConfigureTableViewCellsHelper.configureSeparator(cell, isLastRow: indexPath.row == DayOfTheWeek.allCases.count - 1)
         
         return cell
     }
