+++ Tracker/TypeTrackersVC/Helpers/TableViewHelper.swift
@@ -9,6 +9,7 @@ import UIKit
 
 // MARK: - TableViewHelper
 final class TableViewHelper {
+    
     static func numberOfRows(
         in viewControllerType: TrackerViewControllerType?,
         section: Int,
@@ -36,22 +37,31 @@ final class TableViewHelper {
         dataProvider: TrackerDataProvider?,
         isAddingCategory: Bool,
         selectedCategory: TrackerCategory?,
-        categories: [TrackerCategory]
+        categories: [TrackerCategory],
+        editingCategoryIndex: IndexPath?,
+        viewController: BaseTrackerViewController
     ) -> UITableViewCell {
         switch viewControllerType {
         case .typeTrackers:
-            return configureTypeTrackersCell(at: indexPath, tableView: tableView)
+            return ConfigureTableViewCellsHelper.configureTypeTrackersCell(
+                at: indexPath,
+                tableView: tableView
+            )
         case .creatingTracker, .schedule:
             return UITableViewCell()
         case .category:
             if isAddingCategory {
-                return configureTextViewCell(
+                
+                return ConfigureTableViewCellsHelper.configureTextViewCell(
                     at: indexPath,
                     tableView: tableView,
-                    categories: categories
+                    categories: categories,
+                    viewController: viewController,
+                    editingCategoryIndex: editingCategoryIndex,
+                    isAddingCategory: isAddingCategory
                 )
             } else {
-                return configureCategoryCell(
+                return ConfigureTableViewCellsHelper.configureCategoryCell(
                     at: indexPath,
                     tableView: tableView,
                     dataProvider: dataProvider,
@@ -64,94 +74,6 @@ final class TableViewHelper {
         }
     }
     
-    private static func configureTypeTrackersCell(
-        at indexPath: IndexPath,
-        tableView: UITableView
-    ) -> UITableViewCell {
-        let cell = UITableViewCell()
-        if #available(iOS 14.0, *) {
-            var content = cell.defaultContentConfiguration()
-            if indexPath.section == 0 {
-                content.text = "Привычка"
-            } else if indexPath.section == 1 {
-                content.text = "Нерегулярное событие"
-            }
-            content.textProperties.alignment = .center
-            content.textProperties.color = .ypBackground
-            content.textProperties.font = UIFont.systemFont(
-                ofSize: 16,
-                weight: .medium
-            )
-            cell.contentConfiguration = content
-        } else {
-            if indexPath.section == 0 {
-                cell.textLabel?.text = "Привычка"
-            } else if indexPath.section == 1 {
-                cell.textLabel?.text = "Нерегулярное событие"
-            }
-            cell.textLabel?.textAlignment = .center
-            cell.textLabel?.font = UIFont.systemFont(
-                ofSize: 16,
-                weight: .medium
-            )
-            cell.textLabel?.textColor = .ypBackground
-        }
-        cell.layer.cornerRadius = 16
-        cell.clipsToBounds = true
-        cell.selectionStyle = .none
-        cell.backgroundColor = .ypBlack
-        return cell
-    }
-    
-
-    private static func configureCategoryCell(
-        at indexPath: IndexPath,
-        tableView: UITableView, 
-        dataProvider: TrackerDataProvider?,
-        categories: [TrackerCategory],
-        selectedCategory: TrackerCategory?
-    ) -> UITableViewCell {
-        guard let cell = tableView.dequeueReusableCell(
-            withIdentifier: CategoryCell.reuseIdentifier,
-            for: indexPath
-        ) as? CategoryCell else {
-            return UITableViewCell()
-        }
-        
-        //let category = categories[indexPath.row]
-        guard let itemTitle = dataProvider?.item(at: indexPath.row) else {
-            print("Ошибка: itemTitle отсутствует для индекса \(indexPath.row)")
-            return UITableViewCell()
-        }
-        print("Категория на индексе \(indexPath.row): \(itemTitle)")
-        cell.configure(with: itemTitle)
-        
-        if let selectedCategory = selectedCategory, selectedCategory.title == itemTitle {
-            cell.accessoryType = .checkmark
-        } else {
-            cell.accessoryType = .none
-        }
-        
-//        configureBaseCell(cell, at: indexPath, totalRows: dataProvider?.numberOfItems ?? 0)
-//        configureSeparator(cell, isLastRow: indexPath.row == (dataProvider?.numberOfItems ?? 0) - 1)
-        return cell
-    }
-
-    private static func configureTextViewCell(
-        at indexPath: IndexPath,
-        tableView: UITableView,
-        categories: [TrackerCategory]
-    ) -> UITableViewCell {
-        guard let cell = tableView.dequeueReusableCell(
-            withIdentifier: TextViewCell.reuseIdentifier,
-            for: indexPath
-        ) as? TextViewCell else {
-            return UITableViewCell()
-        }
-        
-        return cell
-    }
-
     static func didSelectRow(
         at indexPath: IndexPath,
         viewControllerType: TrackerViewControllerType?,
@@ -169,5 +91,22 @@ final class TableViewHelper {
         case .none:
             break
         }
+    }
+    
+    static func calculateCellHeight(for tableView: UITableView, itemCount: Int, itemsPerRow: Int) -> CGFloat {
+        let collectionViewWidth = tableView.frame.width
+        let cellSpacing: CGFloat = 5
+        let leftInset: CGFloat = 10
+        let rightInset: CGFloat = 10
+
+        let totalSpacing = (CGFloat(itemsPerRow - 1) * cellSpacing)
+        let totalInsets = leftInset + rightInset
+        let availableWidth = collectionViewWidth - totalInsets - totalSpacing
+        let itemWidth = availableWidth / CGFloat(itemsPerRow)
+        
+        let numberOfRows = ceil(CGFloat(itemCount) / CGFloat(itemsPerRow))
+        let totalHeight = numberOfRows * itemWidth + (numberOfRows - 1) * cellSpacing
+        
+        return totalHeight + 20
     }
 }
